requirements.txt
# Core NL2SQL Dependencies
semantic-kernel==1.34.0
fastmcp
a2a-sdk==0.2.12
python-dotenv
aiohttp
httpx
pydantic

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Optional: for development and testing
pytest
pytest-asyncio
